Update: user utils 파일 분리 및 코드 정리
