+++ src/utils/users/index.ts
@@ -1,0 +1,3 @@
+export { checkPassword, checkEmptyObject } from "./checkDatas";
+export { handleSignup, handleLogin, handleChange } from "./handleActions";
+export { validateEmail, validatePassword } from "./ValidateDatas";
